requirements.txt
pillow
scipy==1.2.0
scikit-image